move base_policy file
