+++ rltoolkit/policy/base_policy.py
@@ -1,8 +1,8 @@
 '''
 Author: jianzhnie@126.com
 Date: 2022-09-01 10:58:47
-LastEditors: jianzhnie@126.com
-LastEditTime: 2022-09-01 12:30:20
+LastEditors: jianzhnie
+LastEditTime: 2022-09-02 19:42:01
 Description:
 
 Copyright (c) 2022 by jianzhnie jianzhnie@126.com, All Rights Reserved.
@@ -12,7 +12,7 @@ import warnings
 
 warnings.simplefilter('default')
 
-from .model import Model, ModelBase
+from ..core.model import Model, ModelBase
 
 __all__ = ['Algorithm']
