+++ rltoolkit/policy/modelfree/dqn.py
@@ -2,7 +2,7 @@
 Author: jianzhnie
 Date: 2022-09-02 12:53:53
 LastEditors: jianzhnie
-LastEditTime: 2022-09-02 14:25:29
+LastEditTime: 2022-09-02 19:43:24
 Description:
 Copyright (c) 2022 by jianzhnie@126.com, All Rights Reserved.
 '''
@@ -12,7 +12,7 @@ import copy
 import torch
 import torch.optim as optim
 
-from rltoolkit.core.algorithm import Algorithm
+from rltoolkit.policy.base_policy import Algorithm
 from rltoolkit.utils.utils import check_model_method
 
 __all__ = ['DQN']
