+++ rltoolkit/policy/multiagent/coma.py
@@ -2,7 +2,7 @@ from copy import deepcopy
 
 import torch
 
-from rltoolkit.core.algorithm import Algorithm
+from rltoolkit.policy.base_policy import Algorithm
 from rltoolkit.utils.utils import check_model_method
 
 __all__ = ['COMA']
